# Speed up `Lab` conversion with `Srgb`/`LinSrgb` fast path, use `palette` `FromStr` impl