--- a/src/bin/kmeans_colors/err.rs
+++ b/src/bin/kmeans_colors/err.rs
@@ -26,13 +26,11 @@ impl From<std::time::SystemTimeError> for CliError {
 
 impl std::fmt::Display for CliError {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match *self {
-            CliError::File(ref err) => write!(f, "File error: {}", err),
-            CliError::InvalidHex => {
-                write!(f, "Error: Invalid hex color length, must be 6 characters.")
-            }
-            CliError::Parse(ref err) => write!(f, "Parse error: {}", err),
-            CliError::Time(ref err) => write!(f, "Time error: {}", err),
+        match self {
+            CliError::File(err) => write!(f, "{err}"),
+            CliError::Parse(err) => write!(f, "{err}"),
+            CliError::Time(err) => write!(f, "{err}"),
+            CliError::InvalidHex => write!(f, "Invalid hex color, must be 3 or 6 digts"),
         }
     }
 }
@@ -41,9 +39,9 @@ impl std::error::Error for CliError {
     fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
         match self {
             CliError::File(err) => Some(err),
-            CliError::InvalidHex => None,
             CliError::Parse(err) => Some(err),
             CliError::Time(err) => Some(err),
+            CliError::InvalidHex => None,
         }
     }
 }