- Swap to using palette's `FromStr` implementation for colors. Now 3 or 6 digit hex colors are permitted.
- Use the special-cased `Srgb`/`LinSrgb` conversion between `u8` and `f32`/`f64` where possible, that version uses a lookup table and approximation

Speedup gains increase as image size increases. On a 4500x3000 image, a 44% speedup was seen reducing the total time to calculate centroids from 7.8s to 4.3s after switching from `into_format` to using `into_linear`. The speedup is closer to 15-17% on smaller images (320x200) but still noticeable when measured.

Example speedups for reading in images in `Rgb`, converting them to `Lab`, and calculating the centroids. Prompted from this discussion https://github.com/Ogeon/palette/discussions/341#discussioncomment-6561118
|Image Size|`into_format`|`into_linear`|
|:-:|-:|-:|
|4547x3032|7.8s|**4.3s**|
|1700x2200|2.4s|**1.3s**|
|805x1000|0.48s|**0.35s**|
|320x206|0.056s|**0.046s**|